pyproject.toml
[project]
name = "activation-patching"
version = "0.1.0"
description = "Activation Patching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers[torch]>=4.53.1",
    "python-dotenv>=1.1.1",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0", 
    "seaborn>=0.12.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
